requirements.txt
Django==1.8.2
django-appconf==1.0.1
flickrapi==2.0
oauthlib==1.0.1
python-dateutil==2.4.2
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
six==1.9.0
twitter==1.17.0
